pyproject.toml
[project]
name = "fastapi-template"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.127.0,<0.128.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlmodel (>=0.0.38,<0.0.39)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "redis (>=7.4.0,<8.0.0)",
    "aiohttp (>=3.13.5,<4.0.0)",
    "aiofiles (>=25.1.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
